updating dev branch
index.js had a load of code that wasnt doing anyting so i deleted it all.

addpost.js was missing the radio tags in the for loop that pushes html into the adpost.ejs file.

there was also a lone div tag in addpost.ejs that looked like it didnt have a closing tag so i deleted that